+++ lib/fields/address.dart
@@ -1,3 +1,4 @@
+import 'package:every_door/fields/helpers/new_addr.dart';
 import 'package:every_door/fields/helpers/radio_field.dart';
 import 'package:every_door/models/address.dart';
 import 'package:every_door/models/amenity.dart';
@@ -50,6 +51,31 @@ class _AddressInputState extends ConsumerState<AddressInput> {
     });
   }
 
+  addAddress(BuildContext context) async {
+    final StreetAddress? addr = await showModalBottomSheet(
+      context: context,
+      isScrollControlled: true,
+      builder: (context) {
+        return SingleChildScrollView(
+          child: Padding(
+            padding: EdgeInsets.only(
+              top: 6.0,
+              left: 10.0,
+              right: 10.0,
+              bottom: MediaQuery.of(context).viewInsets.bottom,
+            ),
+            child: NewAddressPane(),
+          ),
+        );
+      },
+    );
+
+    if (addr == null || addr.isEmpty) return;
+    setState(() {
+      addr.setTags(widget.element);
+    });
+  }
+
   @override
   Widget build(BuildContext context) {
     if (nearestAddresses.isEmpty) {
@@ -58,13 +84,15 @@ class _AddressInputState extends ConsumerState<AddressInput> {
 
     final current = StreetAddress.fromTags(widget.element.getFullTags());
     return RadioField(
-      options: nearestAddresses.map((e) => e.toString()).toList(),
+      options: nearestAddresses.map((e) => e.toString()).toList() + ['+'],
       value: current.isEmpty ? null : current.toString(),
       onChange: (value) {
         if (value == null) {
           setState(() {
             StreetAddress.clearTags(widget.element);
           });
+        } else if (value == '+') {
+          addAddress(context);
         } else {
           final addr = nearestAddresses
               .firstWhere((element) => element.toString() == value);